feat(requests): move requests component back to sm module
Requests are now only used from the sm module

Fixes #332
