+++ src/app/service-management/request-detail/request-detail.component.scss
@@ -1,4 +1,4 @@
-@import "../../../../constants.scss";
+@import "../../../constants.scss";
 app-request-detail {
 	overflow-y: auto;
 	display: flex;
