+++ src/app/service-management/requests/requests.component.html
@@ -1,4 +1,4 @@
-<h3 class="title3 header-section-name">{{ section }}</h3>
+<h3 class="title3 header-section-name" i18n="@@sm">Service management</h3>
 
 <h1 class="title1 title-section-name" i18n="@@requests">Requests</h1>
