+++ src/app/service-management/service-management.module.ts
@@ -18,6 +18,8 @@ import { NetworkServiceInstancesComponent } from './network-service-instances/ne
 import { NetworkServiceInstancesDetailComponent } from './network-service-instances-detail/network-service-instances-detail.component';
 import { VnfRecordDetailComponent } from './vnf-record-detail/vnf-record-detail.component';
 import { CnfRecordDetailComponent } from './cnf-record-detail/cnf-record-detail.component';
+import { RequestsComponent } from './requests/requests.component';
+import { RequestDetailComponent } from './request-detail/request-detail.component';
 import { LicenceListComponent } from './license-list/license-list.component';
 import { LicenceDetailComponent } from './license-detail/license-detail.component';
 
@@ -33,6 +35,8 @@ import { LicenceDetailComponent } from './license-detail/license-detail.componen
 		NetworkServiceInstancesDetailComponent,
 		VnfRecordDetailComponent,
 		CnfRecordDetailComponent,
+		RequestsComponent,
+		RequestDetailComponent,
 		LicenceListComponent,
 		LicenceDetailComponent
 	],
