+++ src/app/shared/services/common/common.service.ts
@@ -9,7 +9,6 @@ import { UtilsService } from './utils.service';
 export class CommonService {
 	authHeaders: HttpHeaders;
 	request_uuid: string;
-	NA = 'Not available';
 
 	constructor(
 		private authService: AuthService,
@@ -289,72 +288,6 @@ export class CommonService {
 					description: response[ 'nsd' ][ 'description' ],
 					vnf: response[ 'nsd' ][ 'network_functions' ]
 				} : { };
-		} catch (error) {
-			console.error(error);
-		}
-	}
-
-	/**
-	 * Retrieves a list of Network Service requests.
-	 * Either following a search pattern or not.
-	 *
-	 * @param search [Optional] Request attributes that must be
-	 *                          matched by the returned list of
-	 *                          NS requests.
-	 */
-	async getRequests(search?) {
-		const headers = this.authService.getAuthHeaders();
-		const url = search !== undefined ?
-			this.config.baseSP + this.config.requests + search :
-			this.config.baseSP + this.config.requests;
-
-		try {
-			const response = await this.http.get(url, { headers: headers }).toPromise();
-			return response instanceof Array ?
-				response.map(item => ({
-					requestId: item.id,
-					name: item.name || 'Unknown',
-					serviceName: item[ 'service' ] ? item.service.name : this.NA,
-					type: item.request_type,
-					createdAt: item.created_at,
-					status: item.status
-				})) : [];
-		} catch (error) {
-			console.error(error);
-		}
-	}
-
-	/**
-	 * Retrieves a Network Service request by UUID
-	 *
-	 * @param uuid UUID of the desired NS request.
-	 */
-	async getOneRequest(uuid: string) {
-		const headers = this.authService.getAuthHeaders();
-		const url = this.config.baseSP + this.config.requests + '/' + uuid;
-
-		try {
-			const response = await this.http.get(url, { headers: headers }).toPromise();
-			return response.hasOwnProperty('id') ?
-				{
-					uuid: response[ 'id' ],
-					name: response[ 'name' ] || 'Unknown',
-					status: response[ 'status' ],
-					type: response[ 'request_type' ],
-					updatedAt: response[ 'updated_at' ],
-					slaUUID: response[ 'sla_id' ],
-					serviceVendor: response[ 'service' ] ?
-						response[ 'service' ][ 'vendor' ] : null,
-					serviceName: response[ 'service' ] ?
-						response[ 'service' ][ 'name' ] : null,
-					serviceVersion: response[ 'service' ] ?
-						response[ 'service' ][ 'version' ] : null,
-					serviceUUID: response[ 'service' ] ?
-						response[ 'service' ][ 'uuid' ] : null,
-					blacklist: response[ 'blacklist' ],
-					ingresses: response[ 'ingresses' ],
-					egresses: response[ 'egresses' ]
-				} : [];
 		} catch (error) {
 			console.error(error);
 		}
